Working UEFI app

+++ Makefile
@@ -11,26 +11,40 @@ OUTFILE=$(OUTPUT_DIR)/KulkarNIX.bin
 # Section : UEFI
 debug: UEFI Kernel MakeUEFI
 	@echo "========== UEFI Debug Image =========="
-	qemu-system-x86_64 -s -S -L /usr/share/ovmf -bios /usr/share/ovmf/OVMF.fd
+	qemu-system-x86_64 -s -S -bios /usr/share/ovmf/OVMF.fd -net none -drive format=raw,unit=0,file=$(OUTPUT_DIR)/UEFI.img
 
 release: UEFI Kernel MakeUEFI
 	@echo "========== UEFI Release Image =========="
-	qemu-system-x86_64 -L /usr/share/ovmf -bios /usr/share/ovmf/OVMF.fd -drive file=$(OUTPUT_DIR)/UEFI.img if=ide
+	qemu-system-x86_64 -bios /usr/share/ovmf/OVMF.fd -net none -drive format=raw,unit=0,file=$(OUTPUT_DIR)/UEFI.img
 
 UEFI:
 	$(MAKE) -C Arch/ $(BOOT)
 
-MakeUEFI:
-	@echo "========== Making UEFI Image =========="
+MakeUEFI: MakeFS MakeMount MakeCP MakeCleanup
+
+MakeFS:
+ifneq ("$(wildcard $(OUTPUT_DIR)/UEFI.img)","")
+	@echo "========== FS image exists, not recreating image =========="
+	sudo losetup --offset 1048576 --sizelimit 46934528 /dev/loop42 $(OUTPUT_DIR)/UEFI.img
+	sudo mkdosfs -F 32 /dev/loop42
+else
 	dd if=/dev/zero of=$(OUTPUT_DIR)/UEFI.img bs=512 count=93750
 	gdisk $(OUTPUT_DIR)/UEFI.img
 	sudo losetup --offset 1048576 --sizelimit 46934528 /dev/loop42 $(OUTPUT_DIR)/UEFI.img
 	sudo mkdosfs -F 32 /dev/loop42
+endif
+
+MakeMount:
 	rm -rf $(OUTPUT_DIR)/FakeMount
 	mkdir $(OUTPUT_DIR)/FakeMount
 	sudo mount /dev/loop42 $(OUTPUT_DIR)/FakeMount
 	sudo mkdir -p $(OUTPUT_DIR)/FakeMount/EFI/BOOT
+
+MakeCP:
 	sudo cp $(OUTPUT_DIR)/BOOTX64.EFI $(OUTPUT_DIR)/FakeMount/EFI/BOOT/
+	sudo cp $(OUTPUT_DIR)/Kernel.bin $(OUTPUT_DIR)/FakeMount/
+
+MakeCleanup:
 	sudo umount $(OUTPUT_DIR)/FakeMount
 	rm -rf $(OUTPUT_DIR)/FakeMount
 	sudo losetup -d /dev/loop42
@@ -65,8 +79,9 @@ Kernel:
 
 # Section : Clean
 clean:
-	rm $(OUTPUT_DIR)/*.sym
-	rm $(OUTPUT_DIR)/*.bin
-	rm $(OUTPUT_DIR)/*.efi
+	-rm $(OUTPUT_DIR)/*.sym
+	-rm $(OUTPUT_DIR)/*.bin
+	-rm $(OUTPUT_DIR)/*.EFI
+	-rm $(OUTPUT_DIR)/*.img
 	$(MAKE) -C Arch/ clean
 	$(MAKE) -C Kernel/ clean
